+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/systems.js
@@ -144,9 +144,8 @@ function gwoSystemChanges() {
           if (system) {
             // Display system description in intelligence panel
             return loc(system[field]()) || "";
-          } else {
-            return "";
           }
+          return "";
         });
       };
 
@@ -313,7 +312,7 @@ function gwoSystemChanges() {
         _.forEach(model.galaxy.systems(), function (system) {
           const star = system.star;
           const ai = star.ai();
-          const guardians = ai.mirrorMode;
+          const guardians = ai && ai.mirrorMode;
 
           if (ai && ai.team === defeatedTeam) {
             const replacementAI = _.first(ai.foes);