+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v5.76.1 - 2025-01-25
+
+### Fixed
+
+- Guardian planet not being pre-dealt tech if you had unlocked all loadouts
+- Cheats not defeating bosses
+- Lucky Commander causing a 404 error if Galactic War Overhaul was uninstalled
+
 ## v5.76.0 - 2024-12-31
 
 ### Added